migrate HTTP file upload to use `mellium.im/xmpp/upload`, refactor slot handling and type definitions

+++ types.go
@@ -126,7 +126,7 @@ type XMPPChatMessage struct {
 
 type HttpUploadComponent struct {
 	Jid         jid.JID
-	MaxFileSize int64
+	MaxFileSize int
 }
 
 type ChatMessageHandler func(client *XmppClient, message *XMPPChatMessage)
